services/time-slot-validation/requirements.txt
paho-mqtt
redis
pillow
pyzbar[scripts]
opencv-python
pytz